Refactor duplication in remote test cases
