+++ integration-test/it-runner/src/test/java/com/example/layrry/integrationtest/AbstractIntegrationTestCase.java
@@ -1,0 +1,50 @@
+/**
+ *  Copyright 2020 The ModiTect authors
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package com.example.layrry.integrationtest;
+
+import org.junit.After;
+import org.junit.Before;
+
+import java.io.ByteArrayOutputStream;
+import java.io.PrintStream;
+
+import static org.junit.Assert.assertTrue;
+
+public abstract class AbstractIntegrationTestCase {
+    private ByteArrayOutputStream sysOut;
+    private PrintStream originalSysOut;
+
+    @Before
+    public void setupSysOut() {
+        sysOut = new ByteArrayOutputStream();
+        System.setOut(new PrintStream(sysOut));
+
+        originalSysOut = System.out;
+    }
+
+    @After
+    public void restoreSysOut() {
+        System.setOut(originalSysOut);
+    }
+
+    protected void assertOutput() {
+        String output = sysOut.toString();
+
+        assertTrue(output.contains("com.example.foo.Foo - Hello, Alice from Foo (Greeter 1.0.0)"));
+        assertTrue(output.contains("com.example.bar.Bar - Hello, Alice from Bar (Greeter 2.0.0)"));
+        assertTrue(output.contains("com.example.bar.Bar - Good bye, Alice from Bar (Greeter 2.0.0)"));
+    }
+}
