+++ integration-test/it-runner/src/test/java/com/example/layrry/integrationtest/AbstractRemoteIntegrationTestCase.java
@@ -1,0 +1,59 @@
+/**
+ *  Copyright 2020 The ModiTect authors
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package com.example.layrry.integrationtest;
+
+import org.eclipse.jetty.server.Handler;
+import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.server.ServerConnector;
+import org.eclipse.jetty.server.handler.DefaultHandler;
+import org.eclipse.jetty.server.handler.HandlerList;
+import org.eclipse.jetty.server.handler.ResourceHandler;
+import org.junit.AfterClass;
+
+import java.nio.file.Path;
+
+public abstract class AbstractRemoteIntegrationTestCase extends AbstractIntegrationTestCase {
+    protected static String serverUri;
+    private static Server server;
+
+    protected static void startServer(ResourceHandler resource_handler) throws Exception {
+        server = new Server();
+        ServerConnector connector = new ServerConnector(server);
+        connector.setPort(0);
+        server.addConnector(connector);
+
+        resource_handler.setDirectoriesListed(true);
+        resource_handler.setResourceBase(Path.of("src", "test", "resources").toAbsolutePath().toString());
+
+        HandlerList handlers = new HandlerList();
+        handlers.setHandlers(new Handler[]{resource_handler, new DefaultHandler()});
+        server.setHandler(handlers);
+
+        server.start();
+
+        String host = connector.getHost();
+        if (host == null) {
+            host = "localhost";
+        }
+        int port = connector.getLocalPort();
+        serverUri = String.format("http://%s:%d/", host, port);
+    }
+
+    @AfterClass
+    public static void stopServer() throws Exception {
+        server.stop();
+    }
+}
