+++ integration-test/it-runner/src/test/java/com/example/layrry/integrationtest/ContentNegotiationRemoteIntegrationTest.java
@@ -15,16 +15,8 @@
  */
 package com.example.layrry.integrationtest;
 
-import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.Request;
-import org.eclipse.jetty.server.Server;
-import org.eclipse.jetty.server.ServerConnector;
-import org.eclipse.jetty.server.handler.DefaultHandler;
-import org.eclipse.jetty.server.handler.HandlerList;
 import org.eclipse.jetty.server.handler.ResourceHandler;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.moditect.layrry.launcher.LayrryLauncher;
@@ -32,46 +24,15 @@ import org.moditect.layrry.launcher.LayrryLauncher;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.PrintStream;
-import java.nio.file.Path;
 
-import static org.junit.Assert.assertTrue;
-
-public class ContentNegotiationRemoteIntegrationTest {
-
-    private static Server server;
-    private static String serverUri;
-    private ByteArrayOutputStream sysOut;
-    private PrintStream originalSysOut;
-
+public class ContentNegotiationRemoteIntegrationTest extends AbstractRemoteIntegrationTestCase{
     @BeforeClass
-    public static void startJetty() throws Exception {
-        server = new Server();
-        ServerConnector connector = new ServerConnector(server);
-        connector.setPort(0);
-        server.addConnector(connector);
-
-        ResourceHandler resource_handler = new MyResourceHandler();
-        resource_handler.setDirectoriesListed(true);
-        resource_handler.setResourceBase(Path.of("src", "test", "resources").toAbsolutePath().toString());
-
-        HandlerList handlers = new HandlerList();
-        handlers.setHandlers(new Handler[]{resource_handler, new DefaultHandler()});
-        server.setHandler(handlers);
-
-        server.start();
-
-        String host = connector.getHost();
-        if (host == null) {
-            host = "localhost";
-        }
-        int port = connector.getLocalPort();
-        serverUri = String.format("http://%s:%d/", host, port);
+    public static void startServer() throws Exception {
+        startServer(new ContentNegotationResourceHandler());
     }
 
-    private static class MyResourceHandler extends ResourceHandler {
+    private static class ContentNegotationResourceHandler extends ResourceHandler {
         @Override
         public void handle(String target, Request baseRequest, HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
             String[] paths = request.getQueryString().split("=");
@@ -80,32 +41,6 @@ public class ContentNegotiationRemoteIntegrationTest {
             response.setHeader("Content-Type", "text/vnd." + ("yml".equals(extension) ? "yaml" : extension));
             super.handle(target, baseRequest, request, response);
         }
-    }
-
-    @AfterClass
-    public static void stopJetty() throws Exception {
-        server.stop();
-    }
-
-    @Before
-    public void setupSysOut() {
-        sysOut = new ByteArrayOutputStream();
-        System.setOut(new PrintStream(sysOut));
-
-        originalSysOut = System.out;
-    }
-
-    @After
-    public void restoreSysOut() {
-        System.setOut(originalSysOut);
-    }
-
-    private void assertOutput() {
-        String output = sysOut.toString();
-
-        assertTrue(output.contains("com.example.foo.Foo - Hello, Alice from Foo (Greeter 1.0.0)"));
-        assertTrue(output.contains("com.example.bar.Bar - Hello, Alice from Bar (Greeter 2.0.0)"));
-        assertTrue(output.contains("com.example.bar.Bar - Good bye, Alice from Bar (Greeter 2.0.0)"));
     }
 
     @Test
